+++ database/migrations/2025_07_15_001300_entrust_setup_tables.php
@@ -31,7 +31,7 @@ class EntrustSetupTables extends Migration
 
         // Create table for associating roles to users (Many-to-Many)
         Schema::create('role_user', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->unsigned();
+            $table->unsignedInteger('user_id')->unsigned();
             $table->integer('role_id')->unsigned();
 
             $table->foreign('user_id')->references('id')->on('users')
