+++ database/migrations/2025_11_25_162830_create_awards_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name', 255);
             $table->text('description')->nullable();
             $table->string('by', 255)->nullable();
-            $table->unsignedBigInteger('vendor_id');
+            $table->unsignedInteger('vendor_id');
             $table->dateTime('created_at');
             $table->dateTime('updated_at');
