+++ database/migrations/2027_02_24_000001_create_jawatankuasas_table.php
@@ -16,11 +16,11 @@ return new class extends Migration
         }
         Schema::create('jawatankuasas', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('tender_id')->nullable();
+            $table->unsignedInteger('tender_id')->nullable();
             $table->enum('jenis_jawatankuasa', ['spec', 'open', 'tech', 'fin'])->index();
             $table->enum('p_p', ['1', '0'])->index();
             $table->enum('peranan', ['1', '2', '3'])->index();
-            $table->unsignedBigInteger('user_id')->nullable();
+            $table->unsignedInteger('user_id')->nullable();
             $table->text('catatan')->nullable();
             $table->string('dokumen_sokongan_nama')->nullable();
             $table->string('dokumen_sokongan_path')->nullable();
