+++ database/migrations/2027_04_27_000006_create_technical_checklist_headers_table.php
@@ -14,15 +14,15 @@ return new class extends Migration
         Schema::create('technical_checklist_headers', function (Blueprint $table) {
             $table->id();
             $table->uuid('uuid')->unique();
-            $table->unsignedBigInteger('tender_id')->unique();
+            $table->unsignedInteger('tender_id')->unique();
             $table->decimal('max_score', 10, 2)->default(0);
             $table->decimal('passing_score', 10, 2)->default(0);
             $table->decimal('passing_percentage', 5, 2)->default(0);
             $table->string('status', 50)->default('draft')->index();
             $table->timestamp('submitted_at')->nullable();
-            $table->unsignedBigInteger('submitted_by')->nullable();
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('submitted_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
             $table->timestamps();
 
             $table->foreign('tender_id', 'tch_tender_fk')
