+++ database/migrations/2027_07_27_000001_create_tender_teknikal_pematuhan_evaluations_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('tender_teknikal_pematuhan_evaluations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('tender_id')->index();
+            $table->unsignedInteger('tender_id')->index();
             $table->unsignedBigInteger('vendor_id')->index();
             $table->uuid('checklist_item_uuid')->index();
             // 1 = Mematuhi (Passed), 0 = Tidak Mematuhi (Failed)
