+++ database/migrations/2027_08_05_000001_create_tender_kewangan_evaluations_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('tender_kewangan_evaluations', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('tender_id')->index();
+            $table->unsignedInteger('tender_id')->index();
             $table->unsignedBigInteger('vendor_id')->index();
             $table->uuid('checklist_item_uuid')->index();
             // 1 = Mematuhi, 0 = Tidak Mematuhi
